flask-api/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
requests==2.31.0
pillow==9.5.0
